// SPA/src/app/interfaces/IUtilizador.ts
export interface IUtilizador{
    id: string;
    email: string;
    password: string;
    nome: string;
    dataNascimento: string;
    estadoEmocional: string;
    dataEstadoEmocional : string;
    codigoPais: string;
    numeroTelemovel:string;
    facebook: string;
    linkedIn: string;
    avatar: string;
    tags: string[];
    descricao: string;
    pais: string;
    cidade:string;
}